+++ sphinxcontrib/osint/plugins/pdf.py
@@ -11,7 +11,7 @@ __author__ = 'bibi21000 aka Sébastien GALLET'
 __email__ = 'bibi21000@gmail.com'
 
 import os
-import logging
+from sphinx.util import logging
 
 from . import reify, PluginSource